finish-commit

+++ manage_user/Module/students/view.php
@@ -17,11 +17,18 @@ layouts('header', $data);
 
 $search = isset($_GET['search']) ? trim($_GET['search']) : ''; // Lấy từ khóa tìm kiếm
 
-$query = "SELECT * FROM students WHERE 1 "; // Truy vấn cơ bản
+$query = "SELECT students.*, user.email AS user_email, user.fullname AS user_fullname 
+          FROM students 
+          LEFT JOIN user ON students.studentId = user.id 
+          WHERE 1 "; // Basic query
 
 if (!empty($search)) {
     // Thêm điều kiện tìm kiếm
-    $query .= "AND (student_name LIKE :search OR student_email LIKE :search OR student_code LIKE :search) ";
+    $query .= "AND (students.student_name LIKE :search 
+                    OR students.student_email LIKE :search 
+                    OR students.student_code LIKE :search 
+                    OR user.fullname LIKE :search 
+                    OR user.email LIKE :search) ";
 }
 
 $query .= "ORDER BY update_at LIMIT :start_from, :per_page_record";
@@ -48,14 +55,15 @@ $smg_type = getFlashData('smg_type');
 // $old = getFlashData('old');
 
 $regexResult = checkPrivilege();
-if (!$regexResult){
-    echo 'Bạn không có quyền truy cập';exit;
+if (!$regexResult) {
+    echo 'Bạn không có quyền truy cập';
+    exit;
 }
 
 ?>
-
 <!DOCTYPE html>
 <html lang="en">
+
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
@@ -66,35 +74,69 @@ if (!$regexResult){
             height: 100vh;
         }
 
-        nav {
-            width: 340px;
-            margin-left: -12px;
-            padding-bottom: 260px;
-        }
-
-        .main-content {
-            flex-grow: 1;
-            padding: 20px;
-        }
-
         .mb {
             padding-bottom: 46rem;
         }
 
-        .strengh {
-            width: 20rem;
-            height: 10rem;
-            margin-top: 3rem;
+        .flex-shrink-0 {
+            height: 100vh;
+            border-right: 1px solid #dee2e6;
         }
 
-        .nav-item:hover {
-            background-color: #343a40;
-            /* Change this to the color you want on hover */
+        .btn-toggle {
+            font-weight: 600;
+            color: #495057;
+            padding: 0.75rem 1rem;
+            width: 100%;
+            text-align: left;
+            font-size: 1.5rem;
         }
 
-        .dp {
-            display: flex;
-            justify-content: space-around;
+        .btn-toggle:hover,
+        .btn-toggle:focus {
+            background-color: #ced4da;
+            color: #212529;
+        }
+
+        .btn-toggle-nav a {
+            padding: 0.5rem 1.5rem;
+            color: #495057;
+            transition: color 0.2s ease;
+            font-size: 1.3rem;
+        }
+
+        .btn-toggle-nav a:hover {
+            color: #007bff;
+        }
+
+        .btn-toggle-nav a.active {
+            background-color: #007bff;
+            color: #fff;
+        }
+
+        .collapse.show {
+            background-color: #f1f3f5;
+        }
+
+        .collapse {
+            padding-left: 0.75rem;
+        }
+
+        .list-unstyled {
+            padding-left: 0;
+        }
+
+        .border-bottom {
+            border-bottom: 1px solid #dee2e6;
+        }
+
+        .link-body-emphasis {
+            color: #212529;
+        }
+
+        .link-body-emphasis:hover {
+            color: #007bff;
+            text-decoration: none;
         }
 
         .pagination {
@@ -122,31 +164,159 @@ if (!$regexResult){
         .table-font {
             font-size: 20px;
         }
+
+        #sidebar.show {
+            display: block;
+        }
+
+        .menu-icon {
+            font-size: 1.8rem;
+            cursor: pointer;
+        }
     </style>
 </head>
+
 <body>
     <div class="container-fluid">
-        <nav class="bg-dark">
-            <ul class="nav nav-item">
-                <li class="nav-item"><a class="nav-link p-4 fs-3 text-white" href="?module=home&action=dashboard"><i class="fa-solid fa-house"><span class="ms-3">DashBoard</span></i></a></li>
+        <div class="menu-icon">
+            <i class="fa-solid fa-bars" id="menu-icon"></i>
+        </div>
+        <div class="flex-shrink-0 p-3 sidebar" id="sidebar" style="width: 280px;">
+            <ul class="list-unstyled ps-0">
+                <li class="mb-1 pb-2">
+                    <button class="btn btn-toggle d-inline-flex align-items-center rounded border-0 collapsed" data-bs-toggle="collapse" data-bs-target="#home-collapse" aria-expanded="false">
+                        Trang home
+                    </button>
+                    <div class="collapse" id="home-collapse" style="">
+                        <ul class="btn-toggle-nav list-unstyled fw-normal pb-1 small">
+                            <li><a href="?module=home&action=dashboard" class="link-body-emphasis d-inline-flex text-decoration-none rounded">Dashboard</a></li>
+                        </ul>
+                    </div>
+                </li>
+                <li class="mb-1 pb-2">
+                    <button class="btn btn-toggle d-inline-flex align-items-center rounded border-0 collapsed" data-bs-toggle="collapse" data-bs-target="#account-collapse" aria-expanded="false">
+                        Quản lý tài khoản
+                    </button>
+                    <div class="collapse" id="account-collapse" style="">
+                        <ul class="btn-toggle-nav list-unstyled fw-normal pb-1 small">
+                            <?php if (checkPrivilege('\?module=user&action=list')) { ?>
+                                <li><a href="?module=user&action=list" class="link-body-emphasis d-inline-flex text-decoration-none rounded">Trang người dùng</a></li>
+                            <?php } ?>
+                            <?php if (checkPrivilege('\?module=user&action=add')) { ?>
+                                <li><a href="?module=user&action=add" class="link-body-emphasis d-inline-flex text-decoration-none rounded">Thêm người dùng</a></li>
+                            <?php } ?>
+                            <?php if (checkPrivilege('\?module=user&action=edit&id=[0-9]*')) { ?>
+                                <li><a href="?module=user&action=edit" class="link-body-emphasis d-inline-flex text-decoration-none rounded">Cập nhật người dùng</a></li>
+                            <?php } ?>
+                            <?php if (checkPrivilege('\?module=user&action=delete&id=[0-9]*')) { ?>
+                                <li><a href="?module=user&action=delete" class="link-body-emphasis d-inline-flex text-decoration-none rounded">Xóa người dùng</a></li>
+                            <?php } ?>
+                        </ul>
+                    </div>
+                </li>
+                <li class="mb-1 pb-2">
+                    <button class="btn btn-toggle d-inline-flex align-items-center rounded border-0 collapsed" data-bs-toggle="collapse" data-bs-target="#student-collapse" aria-expanded="false">
+                        Quản lý sinh viên
+                    </button>
+                    <div class="collapse" id="student-collapse" style="">
+                        <ul class="btn-toggle-nav list-unstyled fw-normal pb-1 small">
+                            <?php if (checkPrivilege('\?module=students&action=view')) { ?>
+                                <li><a href="?module=students&action=view" class="link-body-emphasis d-inline-flex text-decoration-none rounded">Trang sinh viên</a></li>
+                            <?php } ?>
+                            <?php if (checkPrivilege('\?module=students&action=add')) { ?>
+                                <li><a href="?module=students&action=add" class="link-body-emphasis d-inline-flex text-decoration-none rounded">Thêm sinh viên</a></li>
+                            <?php } ?>
+                            <?php if (checkPrivilege('\?module=students&action=update&student_id=[0-9]*')) { ?>
+                                <li><a href="?module=students&action=update" class="link-body-emphasis d-inline-flex text-decoration-none rounded">Cập nhật sinh viên</a></li>
+                            <?php } ?>
+                            <?php if (checkPrivilege('\?module=students&action=delete&student_id=[0-9]*')) { ?>
+                                <li><a href="?module=students&action=delete" class="link-body-emphasis d-inline-flex text-decoration-none rounded">Xóa sinh viên</a></li>
+                            <?php } ?>
+                        </ul>
+                    </div>
+                </li>
+                <li class="mb-1 pb-2">
+                    <button class="btn btn-toggle d-inline-flex align-items-center rounded border-0 collapsed" data-bs-toggle="collapse" data-bs-target="#section-collapse" aria-expanded="false">
+                        Quản lý học phần
+                    </button>
+                    <div class="collapse" id="section-collapse" style="">
+                        <ul class="btn-toggle-nav list-unstyled fw-normal pb-1 small">
+                            <?php if (checkPrivilege('\?module=course&action=view_course')) { ?>
+                                <li><a href="?module=course&action=view_course" class="link-body-emphasis d-inline-flex text-decoration-none rounded">Trang học phần</a></li>
+                            <?php } ?>
+                            <?php if (checkPrivilege('\?module=course&action=add_course')) { ?>
+                                <li><a href="?module=course&action=add_course" class="link-body-emphasis d-inline-flex text-decoration-none rounded">Thêm học phần</a></li>
+                            <?php } ?>
+                            <?php if (checkPrivilege('\?module=course&action=update_course&course_id=[0-9]*')) { ?>
+                                <li><a href="?module=course&action=update_course" class="link-body-emphasis d-inline-flex text-decoration-none rounded">Cập nhật học phần</a></li>
+                            <?php } ?>
+                            <?php if (checkPrivilege('\?module=course&action=delete_course&course_id=[0-9]*')) { ?>
+                                <li><a href="?module=course&action=delete_course" class="l-ink-body-emphasis d-inline-flex text-decoration-none rounded">Xóa học phần</a></li>
+                            <?php } ?>
+                        </ul>
+                    </div>
+                </li>
+                <li class="mb-1 pb-2">
+                    <button class="btn btn-toggle d-inline-flex align-items-center rounded border-0 collapsed" data-bs-toggle="collapse" data-bs-target="#transcript-collapse" aria-expanded="false">
+                        Quản lý bảng điểm
+                    </button>
+                    <div class="collapse" id="transcript-collapse" style="">
+                        <ul class="btn-toggle-nav list-unstyled fw-normal pb-1 small">
+                            <?php if (checkPrivilege('\?module=scoresheets&action=student_courses')) { ?>
+                                <li><a href="?module=scoresheets&action=student_courses" class="link-body-emphasis d-inline-flex text-decoration-none rounded">Trang bảng điểm</a></li>
+                            <?php } ?>
+                            <?php if (checkPrivilege('\?module=scoresheets&action=add_score')) { ?>
+                                <li><a href="?module=scoresheets&action=add_score" class="link-body-emphasis d-inline-flex text-decoration-none rounded">Thêm bảng điểm</a></li>
+                            <?php } ?>
+                            <?php if (checkPrivilege('\?module=scoresheets&action=edit_scoresheets&id=[0-9]*')) { ?>
+                                <li><a href="?module=scoresheets&action=edit_scoresheets" class="link-body-emphasis d-inline-flex text-decoration-none rounded">Cập nhật bảng điểm</a></li>
+                            <?php } ?>
+                            <?php if (checkPrivilege('\?module=scoresheets&action=delete_score&id=[0-9]*')) { ?>
+                                <li><a href="?module=scoresheets&action=delete_score" class="link-body-emphasis d-inline-flex text-decoration-none rounded">Xóa bảng điểm</a></li>
+                            <?php } ?>
+                            <?php if (checkPrivilege('\?module=scoresheets&action=print')) { ?>
+                                <li><a href="?module=scoresheets&action=print" class="link-body-emphasis d-inline-flex text-decoration-none rounded">In bảng điểm</a></li>
+                            <?php } ?>
+                        </ul>
+                    </div>
+                </li>
+                <li class="mb-1">
+                    <button class="btn btn-toggle d-inline-flex align-items-center rounded border-0 collapsed" data-bs-toggle="collapse" data-bs-target="#recognize-collapse" aria-expanded="false">
+                        Kết quả công nhận điểm
+                    </button>
+                    <div class="collapse" id="recognize-collapse" style="">
+                        <ul class="btn-toggle-nav list-unstyled fw-normal pb-1 small">
+                            <?php if (checkPrivilege('\?module=viewdetail&action=view_result')) { ?>
+                                <li><a href="?module=viewdetail&action=view_result" class="link-body-emphasis d-inline-flex text-decoration-none rounded">Trang kết quả công nhận điểm</a></li>
+                            <?php } ?>
+                            <?php if (checkPrivilege('\?module=viewdetail&action=add_result')) { ?>
+                                <li><a href="?module=viewdetail&action=add_result" class="link-body-emphasis d-inline-flex text-decoration-none rounded">Thêm kết quả công nhận điểm</a></li>
+                            <?php } ?>
+                            <?php if (checkPrivilege('\?module=viewdetail&action=edit_result&result_id=[0-9]*')) { ?>
+                                <li><a href="?module=viewdetail&action=edit_result" class="link-body-emphasis d-inline-flex text-decoration-none rounded">Cập nhật kết quả công nhận điểm</a></li>
+                            <?php } ?>
+                            <?php if (checkPrivilege('\?module=viewdetail&action=delete_result&result_id=[0-9]*')) { ?>
+                                <li><a href="?module=viewdetail&action=delete_result" class="link-body-emphasis d-inline-flex text-decoration-none rounded">Xóa kết quả công nhận điểm</a></li>
+                            <?php } ?>
+                        </ul>
+                    </div>
+                </li>
             </ul>
-            <ul class="nav flex-column fs-4">
-                <li class="nav-item mb-3 "><a class="nav-link text-white" href="?module=user&action=list">Quản lý tài khoản</a></li>
-                <li class="nav-item mb-3 "><a class="nav-link text-white" href="?module=students&action=view">Quản lý sinh viên</a></li>
-                <li class="nav-item mb-3 "><a class="nav-link text-white" href="?module=course&action=view_course">Quản lý học phần</a></li>
-                <li class="nav-item mb-3 "><a class="nav-link text-white" href="?module=viewdetail&action=view_result">Quản lý kết quả công nhận điểm</a></li>
-                <li class="nav-item mb-3 "><a class="nav-link text-white" href="?module=scoresheets&action=print">In ấn bảng điểm</a></li>
-                <li class="nav-item mb-3 "><a class="nav-link text-white" href="?module=students&action=filter">Lọc danh sách sinh viên công nhận điểm</a></li>
-            </ul>
-        </nav>
+        </div>
 
         <div class="container">
             <hr>
             <h2>Quản Lý Sinh Viên</h2>
             <div class="d-flex justify-content-between align-items-center mb-3">
-                <!-- Nút thêm sinh viên -->
-                <a href="?module=students&action=add" class="btn btn-success btn-sm fs-5">Thêm sinh viên<i class="fa-solid fa-plus ms-2"></i></a>
-
+                <div>
+                    <!-- Nút thêm sinh viên -->
+                    <?php if (checkPrivilege('\?module=students&action=add')) { ?>
+                        <a href="?module=students&action=add" class="btn btn-success btn-sm fs-5">Thêm sinh viên<i class="fa-solid fa-plus ms-2"></i></a>
+                    <?php } ?>
+                    <?php if (checkPrivilege('\?module=students&action=filter')) { ?>
+                        <a href="?module=students&action=filter" class="btn btn-success btn-sm fs-5">Lọc sinh viên<i class="fa-solid fa-plus ms-2"></i></a>
+                    <?php } ?>
+                </div>
                 <!-- Thanh tìm kiếm -->
                 <form method="GET" action="">
                     <input type="hidden" name="module" value="students">
@@ -171,15 +341,18 @@ if (!$regexResult){
                     <th>Ngày sinh</th>
                     <th>Khoa</th>
                     <th>Khóa học</th>
-                    <th width="5%">Sửa</th>
-                    <th width="5%">Xóa</th>
+                    <?php if (checkPrivilege('\?module=students&action=update&student_id=[0-9]*')) { ?>
+                        <th width="5%">Sửa</th>
+                    <?php } ?>
+                    <?php if (checkPrivilege('\?module=students&action=delete&student_id=[0-9]*')) { ?>
+                        <th width="5%">Xóa</th>
+                    <?php } ?>
                 </thead>
                 <tbody>
                     <?php
                     if (!empty($ListUser)):
                         $count = ($page - 1) * $per_page_record + 1;
                         foreach ($ListUser as $item):
-
                     ?>
                             <tr>
                                 <td><?php echo $count++; ?></td>
@@ -189,8 +362,12 @@ if (!$regexResult){
                                 <td><?php echo $item['date']; ?></td>
                                 <td><?php echo $item['department']; ?></td>
                                 <td><?php echo $item['course']; ?></td>
-                                <td><a href="<?php echo _WEB_HOST; ?>?module=students&action=update&student_id=<?php echo $item['student_id']; ?>" class="btn btn-warning btn-sm"><i class="fa-solid fa-pen-to-square"></i></a></td>
-                                <td><a href="<?php echo _WEB_HOST; ?>?module=students&action=delete&student_id=<?php echo $item['student_id']; ?>" onclick="return confirm('Bạn có chắc muốn xóa?')" class="btn btn-warning btn-sm"><i class="fa-solid fa-trash"></i></a></td>
+                                <?php if (checkPrivilege('\?module=students&action=update&student_id=[0-9]*')) { ?>
+                                    <td><a href="<?php echo _WEB_HOST; ?>?module=students&action=update&student_id=<?php echo $item['student_id']; ?>" class="btn btn-warning btn-sm"><i class="fa-solid fa-pen-to-square"></i></a></td>
+                                <?php } ?>
+                                <?php if (checkPrivilege('\?module=students&action=delete&student_id=[0-9]*')) { ?>
+                                    <td><a href="<?php echo _WEB_HOST; ?>?module=students&action=delete&student_id=<?php echo $item['student_id']; ?>" onclick="return confirm('Bạn có chắc muốn xóa?')" class="btn btn-warning btn-sm"><i class="fa-solid fa-trash"></i></a></td>
+                                <?php } ?>
                             </tr>
                         <?php
                         endforeach;
@@ -246,7 +423,14 @@ if (!$regexResult){
             </div>
         </div>
     </div>
+    <script>
+        document.getElementById('menu-icon').addEventListener('click', function() {
+            var sidebar = document.getElementById('sidebar');
+            sidebar.classList.toggle('collapse');
+        });
+    </script>
 </body>
+
 </html>
 
 <?php
